DES in CBC Mode -- DESCBCCipher.py
This currently does everything in encryption except for converting back from binary to string, and looping through all blocks in the plaintext.  The decryption hasn't been done at all, but it should be pretty easy to reverse the process once the encryption is finished.  I have a lot of hours tomorrow to commit to it.